backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
python-dotenv
requests
pillow
